Get rid of Compressed modules used for arithmetics on the compressed form
This is the first commit of a series getting rid of compressed arithmetic.

The Compressed modules abstract a full arithmetic of curve points in
a compressed form. However, under the hood, it uncompresses the points and
returns back a compressed version of the result. It can simply be abstracted
with functions of_compressed_bytes_[exn/opt] and to_compressed_bytes if
providing a compressed bytes version of the point is required.
These functions will be added in a following commit.