Use caml_alloc_custom_mem instead of caml_alloc_custom for blst_pairing struct
See the doc: https://v2.ocaml.org/manual/intfc.html#s%3Ac-custom
> Use this function when your custom block holds only out-of-heap memory (memory
> allocated with malloc or caml_stat_alloc) and no other resources. used should be
> the number of bytes of out-of-heap memory that are held by your custom block.
> This function works like caml_alloc_custom except that the max parameter is
> under the control of the user (via the custom_major_ratio, custom_minor_ratio,
> and custom_minor_max_size parameters) and proportional to the heap sizes.

It also keeps track of the C allocated memory since 4.12.0, so we can trace it
with memtrace.
Rely on the GC heuristic to trigger a collection. IIUC, when we use used = 0 and
max = 1, the custom block does not interfer with the heuristic. So we could
imagine a lot of blst_pairing values are allocated and the GC would never
run (or not as often as it should) and a signficant amount of garbaged
values (out_of_heap_memory * N where N is the number of blst_pairing values)
would still be alived and not collected until other values trigger a compaction.
out_of_heap_memory_size is around 3kB, which is large compared to the caml
block size.